+++ Sources/SwagGen/GenerateCommand.swift
@@ -1,5 +1,6 @@
 import Foundation
 import PathKit
+import Rainbow
 import SwagGenKit
 import Swagger
 import SwiftCLI
@@ -70,9 +71,9 @@ class GenerateCommand: Command {
                 Path(#file).parent().parent().parent() + relativePath,
             ]
 
-            if let symlink = try? bundlePath.symlinkDestination() {
+            if let symlink = try? (bundlePath + "swaggen").symlinkDestination() {
                 possibleSettingsPaths = [
-                    symlink + relativePath,
+                    symlink.parent() + relativePath,
                 ] + possibleSettingsPaths
             }
 
@@ -109,13 +110,15 @@ class GenerateCommand: Command {
             exitWithError("Error loading Swagger Spec: \(error)")
         }
 
-        let specCounts = getCountString(counts: [
-            ("operation", spec.paths.reduce(0) { $0 + $1.operations.count }),
-            ("definition", spec.definitions.count),
-            // ("tag", spec.tags.count),
-            ("parameter", spec.parameters.count),
-            ("security definition", spec.securityDefinitions.count),
-        ], pluralise: true)
+        let specCounts = getCountString(
+            counts: [
+                ("operation", spec.paths.reduce(0) { $0 + $1.operations.count }),
+                ("definition", spec.definitions.count),
+                // ("tag", spec.tags.count),
+                ("parameter", spec.parameters.count),
+                ("security definition", spec.securityDefinitions.count),
+            ], pluralise: true
+        )
         standardOut("Loaded spec: \"\(spec.info.title)\" - \(specCounts)")
 
         //    let invalidReferences = Array(Set(spec.invalidReferences)).sorted()
@@ -130,11 +133,13 @@ class GenerateCommand: Command {
             exitWithError("Error loading template: \(error)")
         }
 
-        let templateCounts = getCountString(counts: [
-            ("template file", templateConfig.templateFiles.count),
-            ("copied file", templateConfig.copiedFiles.count),
-            ("option", templateConfig.options.keys.count),
-        ], pluralise: true)
+        let templateCounts = getCountString(
+            counts: [
+                ("template file", templateConfig.templateFiles.count),
+                ("copied file", templateConfig.copiedFiles.count),
+                ("option", templateConfig.options.keys.count),
+            ], pluralise: true
+        )
         standardOut("Loaded template: \(templateCounts)")
 
         if verbose.value {