+++ Tests/SwagGenKitTests/XCTest.swift
@@ -3,5 +3,6 @@ import XCTest
 class SwagGenTests: XCTestCase {
     func testSwagGenKit() {
         testFixtures()
+        testGenerator()
     }
 }