+++ CHANGELOG.md
@@ -1,8 +1,91 @@
 # Change Log
+## Master
 
-## [2.1.0](https://github.com/yonaskolb/SwagGen/compare/2.0.0...2.1.0)
+## 3.0.2
+
 ### Added
-- Seperated `date` and `date-time` formats into `DateDay` and `DateTime` structs #74 #77
+- Added `example` and `default` to the generator
+
+### Fixed
+- Changed default date formatter in templates to use `yyyy` not `YYY` #114
+- Fixed date formatting of `DateDay` properties #114
+- Fixed encoding of dictionary types #113
+
+### Internal
+- Updated to Swift 4.2
+- Updated YAMS, Rainbow and SwiftCLI
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/3.0.1...3.0.2)
+
+## 3.0.1
+
+### Added
+- Added new `modelProtocol` template option which defaults to `APIModel` #109
+
+### Fixed
+- Fixed crash in Swift template when not using any RequestBehaviours #108
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/3.0.0...3.0.1)
+
+## 3.0.0
+
+### Added
+- Added File upload support #103
+- Added top level security support #104
+- Added `modelType` option to Swift template for class or struct models #94
+- Added `modelInheritance` template option #94
+- Added `modelNames` and `enumNames` template options for overriding names #95
+- Added `x-enum-name` property to Swagger for custom enum names #98
+- Added operation `summary` to generation and template
+
+### Changed
+- Swift template changes #104
+	- Renamed `APIError` to `APIClientError`
+	- Removed `APIClient.authorizer`
+	- Added `RequestBehavour.validate` (replaces `APIClient.authorizer`)
+	- `APIClient.makeRequest` now returns a `CancellableRequest` instead of `Alamofire.Request`
+	- A new `APIClient.jsonDecoder` property which is used for json requests
+	- Renamed `queue` to `completionQueue` in `APIClient.makeRequest`
+	- Replaced `APIError. authorizationError` with `APIClientError. validationError`
+	- Rename `APIService.authorization` to `APIService.securityRequirement`
+- Generated type changes in Swift template. You will now have to handle or typealias the following types #104
+	- `ID`: The `UUID` format. Usually `UUID` or `String`
+	- `File`: The `file` format. Usually `URL`, `Data` or a custom type with a mimeType and fileName
+
+### Fixed
+- Sort operations in generated Readme
+- Better name camel casing #100
+- Fix empty string field decoding in YAML #101
+- Escape `Protocol` in swift templates
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/2.1.2...3.0.0)
+
+## 2.1.2
+
+### Added
+- Added generated Swift template documenation #87
+
+### Fixed
+- Fixed nil `AnyCodable` values being encoded as null
+- Fixed inbuilt templates not being found in Mint installations
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/2.1.1...2.1.2)
+
+## 2.1.1
+
+### Added
+- Added support for ASCI encoded swagger specs #80
+
+### Fixed
+- Fixed homebrew installation on machines with both Xcode 9.3 and Command line tools installed
+- Fixed `UUID` parameter encoding #81
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/2.1.0...2.1.1)
+
+## 2.1.0
+
+### Added
+- Separated `date` and `date-time` formats into `DateDay` and `DateTime` structs #74 #77
 - Added new `modelPrefix` and `modelSuffix` options #75
 
 ### Fixed
@@ -10,7 +93,10 @@
 - Fixed `safeOptionalDecoding` not working on optional Arrays
 - Fixed `date-time` not decoding in some cases #77
 
-## [2.0.0](https://github.com/yonaskolb/SwagGen/compare/1.2.0...2.0.0)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/2.0.0...2.1.0)
+
+## 2.0.0
+
 ### Added
 - Swift template: added `Codable` support to models #61
 - Swift template: added `Equatable` support to models #63
@@ -41,7 +127,10 @@
 - **BREAKING** Swift template: models no longer have `init(jsonDictionary: JSONDictionary)` or `encode() -> JSONDictionary` functions #61
 - Swift template: removed `JSONUtilities` dependency #61
 
-## [1.2.0](https://github.com/yonaskolb/SwagGen/compare/1.1.0...1.2.0)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/1.2.0...2.0.0)
+
+## 1.2.0
+
 ### Added
 - Added `fixedWidthIntegers` option to Swift template. Thanks @martinknabbe
 - Added support for response references #58
@@ -50,7 +139,10 @@
 - Fixed Swift 4.0.2 warnings
 - Fixed Brew install
 
-## [1.1.0](https://github.com/yonaskolb/SwagGen/compare/1.0.0...1.1.0)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/1.1.0...1.2.0)
+
+## 1.1.0
+
 ### Added
 - Generate typealias for models with reference and array types #42 thanks @Liquidsoul
 - generate typealias for models that only have additional properties
@@ -58,7 +150,10 @@
 ### Fixed
 - fixed SPM installation issue
 
-## [1.0.0](https://github.com/yonaskolb/SwagGen/compare/0.6.1...1.0.0)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/1.0.0...1.1.0)
+
+## 1.0.0
+
 ### Added
 - Swift template: decode response on background queue and then call completion on main thread or new `queue` parameter
 
@@ -66,7 +161,10 @@
 - Updated project to Swift 4 #42
 - Updated Swift templates to Swift 4 #42
 
-## [0.6.1](https://github.com/yonaskolb/SwagGen/compare/0.6.0...0.6.1)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.6.1...1.0.0)
+
+## 0.6.1
+
 ### Added
 - Homebrew and Make installations
 - Enums now also have a raw property to access original json
@@ -77,7 +175,10 @@
 ### Fixed
 - Fixed parameters with a file type not being generated
 
-## [0.6.0](https://github.com/yonaskolb/SwagGen/compare/0.5.3...0.6.0)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.6.0...0.6.1)
+
+## 0.6.0
+
 This includes a large rewrite with a lot more test cases so many more specs should be supported
 
 ### Added
@@ -105,7 +206,10 @@ This includes a large rewrite with a lot more test cases so many more specs shou
 - Within templates `tags` is now just a list of all tag names. The previous tag dictionary which contains `name` and `operations` has been moved to `operationsByTag`
 - request response enum cases have been renamed
 
-## [0.5.3](https://github.com/yonaskolb/SwagGen/compare/0.5.2...0.5.3)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.5.3...0.6.0)
+
+## 0.5.3
+
 ### Swift template fixes
 - fixed not building with Swift Package Manager in certain situations
 - fixed array bodies being generated as inline classes
@@ -123,7 +227,10 @@ This includes a large rewrite with a lot more test cases so many more specs shou
 ### Added
 Added suite of tests for parsing, generating and compiling templates from a list of specs. Will improve stability and help prevent regressions. Still some work to do in this area
 
-## [0.5.2](https://github.com/yonaskolb/SwagGen/compare/0.5.1...0.5.2)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.5.2...0.5.3)
+
+## 0.5.2
+
 ### Added
 - added SuccessType typealias to APIResponseValue. This lets you map from a response to successful value
 
@@ -131,8 +238,10 @@ Added suite of tests for parsing, generating and compiling templates from a list
 - Replaced `CustomDebugStringConvertible` with `PrettyPrinted` conformance on Models, so you can specify your own `CustomDebugStringConvertible`. Same string is available at `model.prettyPrinted`
 - Moved generated request enums and anonymous schema from APIRequest.Request to one level higher in scope
 
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.5.1...0.5.2)
 
-## [0.5.1](https://github.com/yonaskolb/SwagGen/compare/0.5.0...0.5.1)
+## 0.5.1
+
 ### Added
 - A request's response now has a responseResult with either `.success(SuccessValue)` or `.failure(FailureValue)`. This is only generated if there is a single schema type for successes responses and a single schema type for failure responses
 
@@ -143,7 +252,9 @@ Added suite of tests for parsing, generating and compiling templates from a list
 ### Fixed
 - Fixed api name not being replaced in `Decoding.swift` anymore
 
-## [0.5.0](https://github.com/yonaskolb/SwagGen/compare/0.4.1...0.5.0)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.5.0...0.5.1)
+
+## 0.5.0
 
 ### Added
 - `APIClient.makeRequest` now returns an Alamofire `Request` if one was created, so requests can now be cancelled
@@ -179,8 +290,10 @@ Models, Requests, Errors and Responses now have CustomStringConvertible and/or C
 ### Fixed
 - Path parameters are no longer also encoded as url parameters in the request template
 
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.4.1...0.5.0)
 
-## [0.4.1](https://github.com/yonaskolb/SwagGen/compare/0.4.0...0.4.1)
+## 0.4.1
+
 ### Fixed
 Improved the generation of complicated specs:
 
@@ -190,7 +303,10 @@ Improved the generation of complicated specs:
 - better support for deeply nested arrays and dictionaries
 - fixed nested enums
 
-## [0.4.0](https://github.com/yonaskolb/SwagGen/compare/0.3...0.4)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.4.0...0.4.1)
+
+## 0.4.0
+
 ### Added
 - Added generated API Client in Swift template #16
 	- monitoring and modification of requests via request behaviours
@@ -207,19 +323,26 @@ Improved the generation of complicated specs:
 ### Fixed
 - Swift names and types are now escaped with a greater range of swift keywords
 
-## [0.3.0](https://github.com/yonaskolb/SwagGen/compare/0.2...0.3)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.3...0.4)
+
+## 0.3.0
 
 ### Fixed
 - Operations with multiple path variables now properly generate an operationId. #11  Thanks @HSchultjan
 - Operation parameters that contain anonymous schemas (those that don't reference a definition schema but define a schema inline) are now genererated properly as nested structs within the APIRequest #13
 
-## [0.2.0](https://github.com/yonaskolb/SwagGen/compare/0.1...0.2)
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.2...0.3)
+
+## 0.2.0
+
 ### Added
 - `Operation`, `Definition`, `Property` and `Parameter`, now have a `raw` property that can be accessed from templates. This represents the raw data that was in the original spec. This lets you access any custom properties you have in your spec
 
 ### Changed
 - `Property` and `Parameter` have lost their `rawType` and `rawName` properties in favour of the above, so they are now `raw.type` and `raw.name`
 - Upgraded Stencil to [0.9](https://github.com/kylef/Stencil/releases/tag/0.9.0)
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/0.1...0.2)
 
 ## 0.1.0
 - First official release